apns: Remove read_until_close streaming_callback
Provides missing error detection in the read_until_close callback.
Removes streaming_callback since the implementation doesn't actually
support handling chunked response data.

If all the result data can be read at once (which is likely for a small
APNs response), then the read_until_close callback is called, not the
read_until_close streaming_callback (which chunks the data).  Only the
streaming_callback was implemented to report an error, so that's not
helpful.